--- src/Logitar.Cms.Ui/components/Auth/index.ts
@@ -1,3 +1,0 @@
-export { ForgotPasswordForm } from './ForgotPasswordForm';
-export { ResetPasswordForm } from './ResetPasswordForm';
-export { SignInForm } from './SignInForm';
