+++ src/Logitar.Cms.Ui/components/AuthLayout/index.ts
@@ -1,0 +1,1 @@
+export { AuthLayout } from './AuthLayout';
