+++ src/Logitar.Cms.Ui/package.json
@@ -38,6 +38,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.3.4",
     "prettier": "2.8.8",
+    "sass": "^1.62.1",
     "typescript": "^5.0.2",
     "vite": "^4.3.0",
     "vite-tsconfig-paths": "^4.2.0"
