+++ src/Logitar.Cms.Ui/pages/index.ts
@@ -1,0 +1,3 @@
+export { ForgotPasswordPage } from './ForgotPassword';
+export { ResetPasswordPage } from './ResetPassword';
+export { SignInPage } from './SignIn';
